requirements.txt
# Core
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.36
plotly>=5.18.0
requests>=2.31.0
python-dotenv>=1.0.0

# ML
scikit-learn>=1.3.0

# Supabase
supabase>=2.3.0

# Broker SDKs (install the ones you need)
kiteconnect>=5.0.0        # Zerodha
smartapi-python>=1.3.0     # Angel One
pyotp>=2.9.0               # TOTP for Angel One

# Utilities
schedule>=1.2.0
